chore: wip, split farm manager tests
