+++ contracts/farm-manager/tests/integration.rs
@@ -18,1379 +18,6 @@ use crate::common::{MOCK_CONTRACT_ADDR_1, MOCK_CONTRACT_ADDR_2};
 mod common;
 
 #[test]
-fn instantiate_farm_manager() {
-    let mut suite =
-        TestingSuite::default_with_balances(vec![coin(1_000_000_000u128, "uom".to_string())]);
-
-    suite.instantiate_err(
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        Coin {
-            denom: "uom".to_string(),
-            amount: Uint128::new(1_000u128),
-        },
-        0,
-        14,
-        86_400,
-        31_536_000,
-        MONTH_IN_SECONDS,
-        Decimal::percent(10),
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::UnspecifiedConcurrentFarms => {}
-                _ => panic!("Wrong error type, should return ContractError::UnspecifiedConcurrentFarms"),
-            }
-        },
-    ).instantiate_err(
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        Coin {
-            denom: "uom".to_string(),
-            amount: Uint128::new(1_000u128),
-        },
-        1,
-        14,
-        86_400,
-        86_399,
-        MONTH_IN_SECONDS,
-        Decimal::percent(10),
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::InvalidUnlockingRange { .. } => {}
-                _ => panic!("Wrong error type, should return ContractError::InvalidUnlockingRange"),
-            }
-        },
-    ).instantiate_err(
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        Coin {
-            denom: "uom".to_string(),
-            amount: Uint128::new(1_000u128),
-        },
-        1,
-        14,
-        86_400,
-        86_500,
-        MONTH_IN_SECONDS,
-        Decimal::percent(101),
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::InvalidEmergencyUnlockPenalty => {}
-                _ => panic!("Wrong error type, should return ContractError::InvalidEmergencyUnlockPenalty"),
-            }
-        },
-    ).instantiate_err(
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        Coin {
-            denom: "uom".to_string(),
-            amount: Uint128::new(1_000u128),
-        },
-        1,
-        14,
-        86_400,
-        86_500,
-        MONTH_IN_SECONDS - 1,
-        Decimal::percent(101),
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::FarmExpirationTimeInvalid { .. } => {}
-                _ => panic!("Wrong error type, should return ContractError::FarmExpirationTimeInvalid"),
-            }
-        },
-    ).instantiate(
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        MOCK_CONTRACT_ADDR_1.to_string(),
-        Coin {
-            denom: "uom".to_string(),
-            amount: Uint128::new(1_000u128),
-        },
-        7,
-        14,
-        86_400,
-        31_536_000,
-        MONTH_IN_SECONDS,
-        Decimal::percent(10), //10% penalty
-    );
-}
-
-#[test]
-fn create_farms() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let invalid_lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_2}/{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, invalid_lp_denom.clone()),
-    ]);
-    suite.instantiate_default();
-
-    let creator = suite.creator().clone();
-    let other = suite.senders[1].clone();
-    let fee_collector = suite.fee_collector_addr.clone();
-
-    suite
-        .query_current_epoch(|result| {
-            let epoch_response = result.unwrap();
-            assert_eq!(epoch_response.epoch.id, 0u64);
-        })
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    // current epoch is 0
-                    start_epoch: Some(0),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::InvalidEpoch { which } => {
-                        assert_eq!(which, "start")
-                    }
-                    _ => panic!("Wrong error type, should return ContractError::InvalidEpoch"),
-                }
-            },
-        );
-
-    for _ in 0..10 {
-        suite.add_one_epoch();
-    }
-    // current epoch is 10
-
-    // try all misconfigurations when creating a farm
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Default::default(),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::InvalidFarmAmount { .. } => {}
-                    _ => panic!("Wrong error type, should return ContractError::InvalidFarmAmount"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(2_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(2_000, "uusdy")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::FarmFeeMissing => {}
-                    _ => {
-                        panic!("Wrong error type, should return ContractError::FarmFeeMissing")
-                    }
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(5_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(8_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(2_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(2_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(5_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(5_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(25),
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmStartTooFar => {}
-                    _ => panic!("Wrong error type, should return ContractError::FarmStartTooFar"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(8),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmStartTimeAfterEndTime => {}
-                    _ => panic!(
-                        "Wrong error type, should return ContractError::FarmStartTimeAfterEndTime"
-                    ),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(15),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmStartTimeAfterEndTime => {}
-                    _ => panic!(
-                        "Wrong error type, should return ContractError::FarmStartTimeAfterEndTime"
-                    ),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    // current epoch is 10
-                    start_epoch: Some(3),
-                    preliminary_end_epoch: Some(5),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::InvalidEpoch { which } => {
-                        assert_eq!(which, "start")
-                    }
-                    _ => panic!("Wrong error type, should return ContractError::InvalidEpoch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    // current epoch is 10
-                    start_epoch: None,
-                    preliminary_end_epoch: Some(5),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::FarmStartTimeAfterEndTime => {}
-                    _ => panic!(
-                        "Wrong error type, should return ContractError::FarmStartTimeAfterEndTime"
-                    ),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(20),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmStartTimeAfterEndTime => {}
-                    _ => panic!(
-                        "Wrong error type, should return ContractError::FarmStartTimeAfterEndTime"
-                    ),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(30),
-                    preliminary_end_epoch: Some(35),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmStartTooFar => {}
-                    _ => panic!("Wrong error type, should return ContractError::FarmStartTooFar"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: invalid_lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                // trying to create a farm for an invalid lp_denom, i.e. an lp_denom that wasn't created
-                // by the pool manager, should fail
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        );
-
-    // create a farm properly
-    suite
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(10_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(10_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                // should fail, max farms per lp_denom was set to 2 in the instantiate_default
-                // function
-                match err {
-                    ContractError::TooManyFarms { .. } => {}
-                    _ => panic!("Wrong error type, should return ContractError::TooManyFarms"),
-                }
-            },
-        )
-        .query_farms(None, None, None, |result| {
-            let farms_response = result.unwrap();
-            assert_eq!(farms_response.farms.len(), 2);
-        })
-        .query_farms(
-            Some(FarmsBy::Identifier("m-farm_1".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                assert_eq!(farms_response.farms.len(), 1);
-                assert_eq!(
-                    farms_response.farms[0].farm_asset,
-                    Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000),
-                    }
-                );
-            },
-        )
-        .query_farms(
-            Some(FarmsBy::Identifier("f-1".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                assert_eq!(farms_response.farms.len(), 1);
-                assert_eq!(
-                    farms_response.farms[0].farm_asset,
-                    Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(10_000),
-                    }
-                );
-            },
-        )
-        .query_farms(
-            Some(FarmsBy::FarmAsset("uusdy".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                assert_eq!(farms_response.farms.len(), 2);
-            },
-        )
-        .query_farms(
-            Some(FarmsBy::LpDenom(lp_denom.clone())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                assert_eq!(farms_response.farms.len(), 2);
-            },
-        )
-        // two farms were created, therefore the fee collector should have received 2_000 uom
-        .query_balance("uom".to_string(), &fee_collector, |balance| {
-            assert_eq!(balance, Uint128::new(2 * 1_000));
-        });
-}
-
-#[test]
-fn expand_farms() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-    ]);
-
-    let creator = suite.creator();
-    let other = suite.senders[1].clone();
-
-    suite.instantiate_default();
-
-    for _ in 0..10 {
-        suite.add_one_epoch();
-    }
-    // current epoch is 10
-
-    suite
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::Unauthorized => {}
-                    _ => panic!("Wrong error type, should return ContractError::Unauthorized"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(8_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_100u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(4_100, "uusdy")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::InvalidExpansionAmount { .. } => {}
-                    _ => panic!(
-                        "Wrong error type, should return ContractError::InvalidExpansionAmount"
-                    ),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_100u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![], // sending no funds when expanding a farm should fail
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::PaymentError(e) => {
-                        assert_eq!(e, PaymentError::NoFunds {})
-                    }
-                    _ => panic!("Wrong error type, should return ContractError::PaymentError"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_100u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(4_100u128, "uom"), coin(4_100u128, "uusdy")], // sending different funds than the one provided in the params should fail
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::PaymentError(e) => {
-                        assert_eq!(e, PaymentError::MultipleDenoms {})
-                    }
-                    _ => panic!("Wrong error type, should return ContractError::PaymentError"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_100u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(4_100u128, "uom")], // sending different funds than the one provided in the params should fail
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => panic!("Wrong error type, should return ContractError::AssetMismatch"),
-                }
-            },
-        )
-        .query_farms(
-            Some(FarmsBy::Identifier("m-farm_1".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                let farm = farms_response.farms[0].clone();
-                assert_eq!(
-                    farm.farm_asset,
-                    Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000),
-                    }
-                );
-
-                assert_eq!(farm.preliminary_end_epoch, 28);
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(20),
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(5_000u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(5_000u128, "uusdy")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .query_farms(
-            Some(FarmsBy::Identifier("m-farm_1".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                let farm = farms_response.farms[0].clone();
-                assert_eq!(
-                    farm.farm_asset,
-                    Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(9_000),
-                    }
-                );
-
-                assert_eq!(farm.preliminary_end_epoch, 38);
-            },
-        );
-}
-#[test]
-fn cant_expand_farm_too_late() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-    ]);
-
-    let creator = suite.creator();
-
-    suite.instantiate_default();
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(1),
-                    preliminary_end_epoch: Some(2),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(1),
-                    preliminary_end_epoch: Some(3),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_2".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        );
-
-    suite
-        .add_one_epoch()
-        .add_one_epoch()
-        .query_current_epoch(|result| {
-            let epoch_response = result.unwrap();
-            assert_eq!(epoch_response.epoch.id, 2);
-        });
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: Some(28),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: Some("m-farm_1".to_string()),
-                },
-            },
-            vec![coin(8_000, "uusdy")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::FarmAlreadyExpired => {}
-                    _ => {
-                        panic!("Wrong error type, should return ContractError::FarmAlreadyExpired")
-                    }
-                }
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: Some(4),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(2_000u128),
-                    },
-                    farm_identifier: Some("m-farm_2".to_string()),
-                },
-            },
-            vec![coin(2_000, "uusdy")],
-            |result| {
-                result.unwrap();
-            },
-        );
-
-    suite
-        .add_one_epoch()
-        .add_one_epoch()
-        .query_current_epoch(|result| {
-            let epoch_response = result.unwrap();
-            assert_eq!(epoch_response.epoch.id, 4);
-        });
-
-    suite.manage_farm(
-        &creator,
-        FarmAction::Fill {
-            params: FarmParams {
-                lp_denom: lp_denom.clone(),
-                start_epoch: None,
-                preliminary_end_epoch: None,
-                curve: None,
-                farm_asset: Coin {
-                    denom: "uusdy".to_string(),
-                    amount: Uint128::new(2_000u128),
-                },
-                farm_identifier: Some("m-farm_2".to_string()),
-            },
-        },
-        vec![coin(2_000u128, "uusdy")],
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::FarmAlreadyExpired => {}
-                _ => panic!("Wrong error type, should return ContractError::FarmAlreadyExpired"),
-            }
-        },
-    );
-}
-
-#[test]
-#[allow(clippy::inconsistent_digit_grouping)]
-fn close_farms() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let lp_denom_2 = format!("factory/{MOCK_CONTRACT_ADDR_1}/2.{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, lp_denom_2.clone()),
-    ]);
-
-    suite.instantiate_default();
-
-    let other = suite.senders[1].clone();
-    let another = suite.senders[2].clone();
-
-    for _ in 0..10 {
-        suite.add_one_epoch();
-    }
-    // current epoch is 10
-
-    suite.manage_farm(
-        &other,
-        FarmAction::Fill {
-            params: FarmParams {
-                lp_denom: lp_denom.clone(),
-                start_epoch: Some(20),
-                preliminary_end_epoch: Some(28),
-                curve: None,
-                farm_asset: Coin {
-                    denom: "uusdy".to_string(),
-                    amount: Uint128::new(4_000u128),
-                },
-                farm_identifier: Some("farm_1".to_string()),
-            },
-        },
-        vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-        |result| {
-            result.unwrap();
-        },
-    );
-    suite
-        .manage_farm(
-            &other,
-            FarmAction::Close {
-                farm_identifier: "m-farm_1".to_string(),
-            },
-            vec![coin(1_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::PaymentError { .. } => {}
-                    _ => panic!("Wrong error type, should return ContractError::PaymentError"),
-                }
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Close {
-                farm_identifier: "m-farm_2".to_string(),
-            },
-            vec![],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::NonExistentFarm => {}
-                    _ => panic!("Wrong error type, should return ContractError::NonExistentFarm"),
-                }
-            },
-        )
-        .manage_farm(
-            &another,
-            FarmAction::Close {
-                farm_identifier: "m-farm_1".to_string(),
-            },
-            vec![],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-                match err {
-                    ContractError::Unauthorized => {}
-                    _ => panic!("Wrong error type, should return ContractError::Unauthorized"),
-                }
-            },
-        )
-        .query_balance("uusdy".to_string(), &other, |balance| {
-            assert_eq!(balance, Uint128::new(999_996_000));
-        })
-        .manage_farm(
-            &other,
-            FarmAction::Close {
-                farm_identifier: "m-farm_1".to_string(),
-            },
-            vec![],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .query_balance("uusdy".to_string(), &other, |balance| {
-            assert_eq!(balance, Uint128::new(1000_000_000));
-        });
-
-    // open new farm
-    suite
-        .query_current_epoch(|result| {
-            let epoch_response = result.unwrap();
-            assert_eq!(epoch_response.epoch.id, 10);
-        })
-        .manage_position(
-            &another,
-            PositionAction::Create {
-                identifier: None,
-                unlocking_duration: 86_400,
-                receiver: None,
-            },
-            vec![coin(1_000, lp_denom_2.clone())],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom_2.clone(),
-                    start_epoch: Some(12),
-                    preliminary_end_epoch: Some(13),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_x".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        );
-
-    for _ in 0..=2 {
-        suite.add_one_epoch();
-    }
-
-    suite.query_current_epoch(|result| {
-        let epoch_response = result.unwrap();
-        assert_eq!(epoch_response.epoch.id, 13);
-    });
-
-    suite
-        .query_balance("uusdy".to_string(), &another, |balance| {
-            assert_eq!(balance, Uint128::new(1_000_000_000));
-        })
-        .claim(&another, vec![], None, |result| {
-            result.unwrap();
-        })
-        .query_farms(
-            Some(FarmsBy::Identifier("m-farm_x".to_string())),
-            None,
-            None,
-            |result| {
-                let farms_response = result.unwrap();
-                let farm = farms_response.farms[0].clone();
-                assert_eq!(
-                    farm.farm_asset,
-                    Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000),
-                    }
-                );
-                // the farm is empty
-                assert_eq!(farm.claimed_amount, Uint128::new(4_000),);
-
-                assert_eq!(farm.preliminary_end_epoch, 13);
-                assert_eq!(farm.start_epoch, 12);
-            },
-        )
-        .query_balance("uusdy".to_string(), &another, |balance| {
-            assert_eq!(balance, Uint128::new(1_000_004_000));
-        })
-        .manage_farm(
-            &other,
-            FarmAction::Close {
-                farm_identifier: "m-farm_x".to_string(),
-            },
-            vec![],
-            |result| {
-                result.unwrap();
-            },
-        );
-}
-
-/// This test recreates the scenario where a malicious TF token freezes token transfers via hooks,
-/// which would brick the rewards claiming mechanism and prevent closing the farm (in case the contract owner
-/// would like to salvage the contract).
-#[test]
-#[allow(clippy::inconsistent_digit_grouping)]
-fn close_farms_wont_fail_with_malicious_tf_token() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let lp_denom_2 = format!("factory/{MOCK_CONTRACT_ADDR_1}/2.{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, lp_denom_2.clone()),
-    ]);
-
-    suite.instantiate_default();
-
-    let other = suite.senders[1].clone();
-
-    // create two farms
-    suite
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &other,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom_2.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uosmo".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm_2".to_string()),
-                },
-            },
-            vec![coin(4_000, "uosmo"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        );
-
-    let farm_manager = suite.farm_manager_addr.clone();
-
-    suite
-        .query_balance("uusdy".to_string(), &farm_manager, |balance| {
-            assert_eq!(balance, Uint128::new(4_000));
-        })
-        .query_balance("uosmo".to_string(), &farm_manager, |balance| {
-            assert_eq!(balance, Uint128::new(4_000));
-        });
-
-    // let's burn tokens from the contract to simulate the case where a malicious TF token freezes
-    // token transfers
-    suite
-        .burn_tokens(&farm_manager, coins(1_000, "uosmo"), |result| {
-            result.unwrap();
-        })
-        .query_balance("uosmo".to_string(), &farm_manager, |balance| {
-            assert_eq!(balance, Uint128::new(3_000));
-        });
-
-    // closing the farm would have failed and bricked the rewards claiming, but not anymore
-    suite.manage_farm(
-        &other,
-        FarmAction::Close {
-            farm_identifier: "m-farm_2".to_string(),
-        },
-        vec![],
-        |result| {
-            assert!(result.unwrap().events.iter().any(|event| {
-                event
-                    .attributes
-                    .iter()
-                    .any(|attr| attr.key == "reason" && !attr.value.is_empty())
-            }));
-        },
-    );
-
-    suite.query_farms(
-        Some(FarmsBy::Identifier("m-farm_2".to_string())),
-        None,
-        None,
-        |result| {
-            let err = result.unwrap_err();
-            assert!(err.to_string().contains("Farm doesn't exist"));
-        },
-    );
-
-    suite.query_farms(None, None, None, |result| {
-        let farms_response = result.unwrap();
-        assert_eq!(farms_response.farms.len(), 1);
-        assert_eq!(farms_response.farms[0].identifier, "m-farm_1");
-    });
-}
-
-#[test]
 fn verify_ownership() {
     let mut suite = TestingSuite::default_with_balances(vec![]);
     let creator = suite.creator();
@@ -4670,120 +3297,6 @@ fn emergency_withdrawal_shares_penalty_with_active_farm_owners() {
         });
 }
 
-#[test]
-fn test_farm_helper() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-    ]);
-
-    let creator = suite.creator();
-    let other = suite.senders[1].clone();
-
-    suite.instantiate_default();
-
-    let farm_manager = suite.farm_manager_addr.clone();
-    let fee_collector = suite.fee_collector_addr.clone();
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("farm".to_string()),
-                },
-            },
-            vec![coin(3_000, "uom")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => {
-                        panic!("Wrong error type, should return ContractError::AssetMismatch")
-                    }
-                }
-            },
-        )
-        .query_balance("uom".to_string(), &creator, |balance| {
-            assert_eq!(balance, Uint128::new(1_000_000_000));
-        })
-        .query_balance("uom".to_string(), &fee_collector, |balance| {
-            assert_eq!(balance, Uint128::zero());
-        })
-        .query_balance("uom".to_string(), &farm_manager, |balance| {
-            assert_eq!(balance, Uint128::zero());
-        })
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(2_000u128),
-                    },
-                    farm_identifier: Some("farm".to_string()),
-                },
-            },
-            vec![coin(2_000, "uusdy"), coin(3_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .query_balance("uom".to_string(), &fee_collector, |balance| {
-            assert_eq!(balance, Uint128::new(1_000));
-        })
-        .query_balance("uom".to_string(), &farm_manager, |balance| {
-            assert_eq!(balance, Uint128::zero());
-        })
-        .query_balance("uom".to_string(), &creator, |balance| {
-            // got the excess of whale back
-            assert_eq!(balance, Uint128::new(999_999_000));
-        });
-
-    suite.manage_farm(
-        &other,
-        FarmAction::Fill {
-            params: FarmParams {
-                lp_denom: lp_denom.clone(),
-                start_epoch: None,
-                preliminary_end_epoch: None,
-                curve: None,
-                farm_asset: Coin {
-                    denom: "uusdy".to_string(),
-                    amount: Uint128::new(2_000u128),
-                },
-                farm_identifier: Some("underpaid_farm".to_string()),
-            },
-        },
-        vec![coin(2_000, "uusdy"), coin(500, "uom")],
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-            match err {
-                ContractError::FarmFeeNotPaid { .. } => {}
-                _ => {
-                    panic!("Wrong error type, should return ContractError::FarmFeeNotPaid")
-                }
-            }
-        },
-    );
-}
-
 /// Complex test case with 4 farms for 2 different LPs somewhat overlapping in time
 /// Farm 1 -> runs from epoch 12 to 16
 /// Farm 2 -> run from epoch 14 to 25
@@ -7480,226 +5993,6 @@ fn test_farm_and_position_id_validation() {
 }
 
 #[test]
-fn fails_to_create_farm_if_more_tokens_than_needed_were_sent() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom"),
-        coin(1_000_000_000u128, "uusdy"),
-        coin(1_000_000_000u128, "uosmo"),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, "invalid_lp"),
-    ]);
-    let creator = suite.creator();
-
-    suite.instantiate_default();
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![
-                coin(4_000, "uusdy"),
-                coin(1_000, "uom"),
-                coin(1_000, "uosmo"),
-            ],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => {
-                        panic!("Wrong error type, should return ContractError::AssetMismatch")
-                    }
-                }
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(5_000, "uom"), coin(1_000, "uosmo")],
-            |result| {
-                let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-                match err {
-                    ContractError::AssetMismatch => {}
-                    _ => {
-                        panic!("Wrong error type, should return ContractError::AssetMismatch")
-                    }
-                }
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(12),
-                    preliminary_end_epoch: Some(16),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(8_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(12),
-                    preliminary_end_epoch: Some(16),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(9_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        );
-}
-
-#[test]
-fn fails_to_create_farm_if_start_epoch_is_zero() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom"),
-        coin(1_000_000_000u128, "uusdy"),
-        coin(1_000_000_000u128, "uosmo"),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, "invalid_lp"),
-    ]);
-    let creator = suite.creator();
-
-    suite.instantiate_default();
-
-    suite.manage_farm(
-        &creator,
-        FarmAction::Fill {
-            params: FarmParams {
-                lp_denom: lp_denom.clone(),
-                start_epoch: Some(0),
-                preliminary_end_epoch: Some(28),
-                curve: None,
-                farm_asset: Coin {
-                    denom: "uusdy".to_string(),
-                    amount: Uint128::new(4_000u128),
-                },
-                farm_identifier: Some("farm_1".to_string()),
-            },
-        },
-        vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-            match err {
-                ContractError::InvalidEpoch { which } => {
-                    assert_eq!(which, "start".to_string())
-                }
-                _ => {
-                    panic!("Wrong error type, should return ContractError::InvalidEpoch")
-                }
-            }
-        },
-    );
-}
-
-#[test]
-fn overriding_farm_with_bogus_id_not_possible() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom"),
-        coin(1_000_000_000u128, "uusdy"),
-        coin(1_000_000_000u128, "uosmo"),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, "invalid_lp"),
-    ]);
-    let creator = suite.creator();
-
-    suite.instantiate_default();
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: Some("1".to_string()),
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: None,
-                    preliminary_end_epoch: None,
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uusdy".to_string(),
-                        amount: Uint128::new(4_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .query_farms(None, None, None, |result| {
-            let farms_response = result.unwrap();
-            assert_eq!(farms_response.farms.len(), 2);
-            assert_eq!(farms_response.farms[0].identifier, "f-1");
-            assert_eq!(farms_response.farms[1].identifier, "m-1");
-        });
-}
-
-#[test]
 fn closing_expired_farm_wont_pay_penalty() {
     let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
     let mut suite = TestingSuite::default_with_balances(vec![
@@ -7841,69 +6134,6 @@ fn providing_custom_position_id_doesnt_increment_position_counter() {
                 assert_eq!(response.positions[2].identifier, "u-custom_id_2");
             },
         );
-}
-
-#[test]
-fn providing_custom_farm_id_doesnt_increment_farm_counter() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom"),
-        coin(1_000_000_000u128, "uusdy"),
-        coin(1_000_000_000u128, "uosmo"),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, "invalid_lp"),
-    ]);
-    let creator = suite.creator();
-
-    suite.instantiate_default();
-
-    suite
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(12),
-                    preliminary_end_epoch: Some(16),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: Some("custom_id_1".to_string()),
-                },
-            },
-            vec![coin(9_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .manage_farm(
-            &creator,
-            FarmAction::Fill {
-                params: FarmParams {
-                    lp_denom: lp_denom.clone(),
-                    start_epoch: Some(12),
-                    preliminary_end_epoch: Some(16),
-                    curve: None,
-                    farm_asset: Coin {
-                        denom: "uom".to_string(),
-                        amount: Uint128::new(8_000u128),
-                    },
-                    farm_identifier: None,
-                },
-            },
-            vec![coin(9_000, "uom")],
-            |result| {
-                result.unwrap();
-            },
-        )
-        .query_farms(None, None, None, |result| {
-            let response = result.unwrap();
-            assert_eq!(response.farms.len(), 2);
-            assert_eq!(response.farms[0].identifier, "f-1");
-            assert_eq!(response.farms[1].identifier, "m-custom_id_1");
-        });
 }
 
 // This is to cover for the following edge case:
@@ -9318,57 +7548,6 @@ fn farm_owners_get_penalty_fees() {
         .query_balance(lp_denom_1.clone(), &bob, |balance| {
             assert_eq!(balance, Uint128::new(1_000_000_000u128 + 25u128));
         });
-}
-
-#[test]
-fn farm_cant_be_created_in_the_past() {
-    let lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_1}/{LP_SYMBOL}").to_string();
-    let invalid_lp_denom = format!("factory/{MOCK_CONTRACT_ADDR_2}/{LP_SYMBOL}").to_string();
-
-    let mut suite = TestingSuite::default_with_balances(vec![
-        coin(1_000_000_000u128, "uom".to_string()),
-        coin(1_000_000_000u128, "uusdy".to_string()),
-        coin(1_000_000_000u128, "uosmo".to_string()),
-        coin(1_000_000_000u128, lp_denom.clone()),
-        coin(1_000_000_000u128, invalid_lp_denom.clone()),
-    ]);
-    suite.instantiate_default();
-
-    let other = suite.senders[1].clone();
-
-    for _ in 0..10 {
-        suite.add_one_epoch();
-    }
-    // current epoch is 10
-
-    // We can create a farm in a past epoch
-    suite.manage_farm(
-        &other,
-        FarmAction::Fill {
-            params: FarmParams {
-                lp_denom: lp_denom.clone(),
-                start_epoch: Some(1), // start epoch in the past
-                preliminary_end_epoch: Some(28),
-                curve: None,
-                farm_asset: Coin {
-                    denom: "uusdy".to_string(),
-                    amount: Uint128::new(4_000u128),
-                },
-                farm_identifier: Some("farm_1".to_string()),
-            },
-        },
-        vec![coin(4_000, "uusdy"), coin(1_000, "uom")],
-        |result| {
-            let err = result.unwrap_err().downcast::<ContractError>().unwrap();
-
-            match err {
-                ContractError::InvalidEpoch { which } => {
-                    assert_eq!(which, "start")
-                }
-                _ => panic!("Wrong error type, should return ContractError::InvalidEpoch"),
-            }
-        },
-    );
 }
 
 /*
